packages/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
h11==0.16.0
hatchling==1.27.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.13.2
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-order==1.3.0
python-dotenv==1.1.1
python-multipart==0.0.20
pyucc==1.7
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
sentry-sdk==2.34.1
shellingham==1.5.4
siblink==1.2.0
sniffio==1.3.1
starlette==0.47.2
trove-classifiers==2025.8.6.13
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
